week_9/ai-agents-mini-bootcamp-main/building-agents/requirements.txt
openai==1.97.1
requests==2.32.4
python-dotenv
pymongo